+++ sybil/parsers/markdown/skip.py
@@ -7,7 +7,7 @@ class SkipParser(AbstractSkipParser):
     A :any:`Parser` for :ref:`skip <markdown-skip-parser>` instructions.
     """
 
-    def __init__(self) -> None:
+    def __init__(self, directive: str = 'skip') -> None:
         super().__init__([
-            DirectiveInHTMLCommentLexer('skip'),
+            DirectiveInHTMLCommentLexer(directive=directive),
         ])