+++ sybil/parsers/rest/skip.py
@@ -7,5 +7,5 @@ class SkipParser(AbstractSkipParser):
     A :any:`Parser` for :ref:`skip <skip-parser>` instructions.
     """
 
-    def __init__(self) -> None:
-        super().__init__([DirectiveInCommentLexer('skip')])
+    def __init__(self, directive: str = 'skip') -> None:
+        super().__init__([DirectiveInCommentLexer(directive=directive)])